Moving Matrix3x3 and Vector3 from framework/mesh to framework/math.
