+++ framework/math/quadratures/angular/angular_quadrature.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "framework/mesh/mesh_vector.h"
+#include "framework/math/vector3.h"
 #include <memory>
 #include <vector>
