+++ framework/math/spatial_discretization/cell_mappings/cell_mapping.cc
@@ -4,7 +4,7 @@
 #include "framework/math/spatial_discretization/cell_mappings/cell_mapping.h"
 #include "framework/math/spatial_discretization/finite_element/finite_element_data.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
-#include "framework/mesh/mesh_matrix3x3.h"
+#include "framework/math/matrix3x3.h"
 #include <utility>
 
 namespace opensn
