+++ framework/math/spatial_discretization/cell_mappings/finite_element/piecewise_linear/piecewise_linear_polyhedron_mapping.h
@@ -6,7 +6,7 @@
 #include "framework/math/spatial_discretization/cell_mappings/finite_element/piecewise_linear/piecewise_linear_base_mapping.h"
 #include "framework/math/quadratures/spatial/tetrahedra_quadrature.h"
 #include "framework/math/quadratures/spatial/triangle_quadrature.h"
-#include "framework/mesh/mesh_matrix3x3.h"
+#include "framework/math/matrix3x3.h"
 
 namespace opensn
 {
