+++ framework/math/vector3.h
