+++ framework/mesh/cell/cell.cc
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
 // SPDX-License-Identifier: MIT
 
-#include "framework/mesh/cell/cell.h"
-#include "framework/mesh/mesh_matrix3x3.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
+#include "framework/mesh/cell/cell.h"
+#include "framework/math/matrix3x3.h"
 #include "framework/data_types/byte_array.h"
 #include "framework/logging/log.h"
 #include "framework/runtime.h"
