+++ framework/mesh/mesh_continuum/mesh_continuum_vertex_handler.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "framework/mesh/mesh_vector.h"
+#include "framework/math/vector3.h"
 #include <map>
 
 namespace opensn
