+++ framework/mesh/mesh_face.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "framework/mesh/mesh_vector.h"
+#include "framework/math/vector3.h"
 #include <array>
 
 namespace opensn
