+++ framework/mesh/raytrace/raytracer.h
@@ -5,7 +5,7 @@
 
 #include "framework/mesh/mesh.h"
 #include "framework/mesh/cell/cell.h"
-#include "framework/mesh/mesh_vector.h"
+#include "framework/math/vector3.h"
 
 namespace opensn
 {
